src/main/java/com/anf/gdi/es/loyalty/consumer/sales/config/OpenApiConfiguration.java
package com.anf.gdi.es.loyalty.consumer.sales.config;

import org.springdoc.core.SpringDocConfigProperties;
import org.springdoc.core.SpringDocConfiguration;
import org.springframework.context.annotation.Bean;
import org.springframework.context.annotation.Configuration;
import org.springframework.context.annotation.Primary;

/**
 * This is the configuration class for OpenAPI configuration.
 */
@Configuration
public class OpenApiConfiguration {

    @Primary
    @Bean
    SpringDocConfiguration springDocConfiguration() {
        return new SpringDocConfiguration();
    }

    @Bean
    public SpringDocConfigProperties springDocConfigProperties() {
        return new SpringDocConfigProperties();
    }
}